replace unnecessary mechanism in e2e home.go w/ociauth fix
